🚚 Rename: 쿼리 프로바이더 위치 변경
rename: 쿼리 프로바이더 위치 변경
