+++ app/main/layout.tsx
@@ -1,4 +1,4 @@
-import Providers from "@/app/_queryFactory/providers";
+import Providers from "@/app/_context/queryProviders";
 import type { Metadata } from "next";
 import "@/app/_styles/globals.css";
