+++ app/profile/[userId]/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { PropsWithChildren } from "react";
-import Providers from "@/app/_queryFactory/providers";
+import Providers from "@/app/_context/queryProviders";
 
 function OnlyHeader({ children }: PropsWithChildren) {
   return <Providers>{children}</Providers>;
