+++ app/profile/[userId]/page.tsx
@@ -1,5 +1,5 @@
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
-import getQueryClient from "@/app/_queryFactory/getQueryClient";
+import getQueryClient from "@/app/_utils/getQueryClient";
 import MyPageContent from "./_components/MyPageContent";
 
 function MyPage() {
